fix: usage of klass->allow_autoconnect_on_external is not right
This needs a different solution. nm_device_autoconnect_allowed() cannot
force autoconnect to happen, it can only prevent it. There is only one caller
of this, NMPolicy.

That one already does:

    if (nm_device_get_act_request(device))
        return;

    if (!nm_device_autoconnect_allowed(device))
        return;

meaning, it won't autoconnect if we have an activation request
already.

However, the entire point of the exercies is to allow autoconnect,
even if the device has an (external) activation request.