fix: dont return NM_DEVICE_CAP_IS_SOFTWARE for loopback
You cannot create or delete loopback devices. They are always there.
Sure, they are purely in software, but still.

If we return that loopback has NM_DEVICE_CAP_IS_SOFTWARE, there are
probably other places that needs to be fixed, where we need to special
case loopback for this reason.

It might be easier to just claim that loopback is not a software device.