device: restart DHCP when the MAC changes and the client-id is fixed
If the MAC changes there is the possibility that the DHCP client will
not be able to renew the address because it uses the old MAC as
CHADDR. Depending on the implementation, the DHCP server might use
CHADDR (so, the old address) as the destination MAC for DHCP replies,
and those packets will be lost.

To avoid this problem, we need to restart the DHCP client after a MAC
change; that has the disadvantage that the server will give out a
different address, unless the client is sending a explicit
client-id. So, limit the restart to situations where the client-id is
explicitly set to a fixed value.

https://bugzilla.redhat.com/show_bug.cgi?id=2110000