team: pass -N twice to teamd
With new teamd versions, this will indicate that we want the team ports to
be kept in place when teamd terminates. Old versions will ignore it and
behave like they always did.

This keeps the team operational when teamd drops off for a while e.g. on
transition from initrd to real root during network boot.